Factor catalog server routes into separate files
