+++ packages/catalog-server/data/README.md
@@ -1,1 +1,1 @@
-`bootstrap-packages.json` has a list of packages known to contain somewhat-valid custom elements monifests. It's used to run an initial import of packages for the catalog.
+`bootstrap-packages.json` has a list of packages known to contain somewhat-valid custom elements manifests. It's used to run an initial import of packages for the catalog.
